Add multiple variants
